tying components to their buttons
